refactor: extract item segments recommendations from items
